backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Data & Validation
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP Client
httpx==0.26.0

# Geospatial
shapely==2.0.2
pyproj==3.6.1

# PDF Generation
weasyprint==60.2

# Utilities
pyyaml==6.0.1
python-dateutil==2.8.2

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
